requirements.txt
gymnasium
stable-baselines3
PyOpenGL
torch
numpy